+++ src/assets/scripts/aircraft/AircraftController.js
@@ -33,24 +33,6 @@ export default class AircraftController {
 
     /**
      * @for AircraftController
-     * @method aircraft_generate_callsign
-     * @param airlineName
-     */
-    aircraft_generate_callsign(airlineName) {
-        // TODO: this should live in the AirlineModel
-        const airline = window.airlineController.airline_get(airlineName);
-
-        if (!airline) {
-            console.warn(`Airline not found: ${airlineName}`);
-
-            return `airline-${airlineName}-not-found`;
-        }
-
-        return airline.generateFlightNumber();
-    }
-
-    /**
-     * @for AircraftController
      * @method aircraft_auto_toggle
      */
     aircraft_auto_toggle() {
@@ -59,19 +41,31 @@ export default class AircraftController {
 
     /**
      * @for AircraftController
-     * @method aircraft_callsign_new
-     * @param airline {string}
+     * @method isCallsignInList
+     * @param callsign {string}
+     * return {boolean}
      */
-    aircraft_callsign_new(airline) {
-        // TODO: the logic needs work here. if `callsign` is always initialized as null, one would imagine that
-        // this function would always result in the creation of a callsign?
-        let callsign = this.aircraft_generate_callsign(airline);
+    isCallsignInList(callsign) {
+        return this.aircraft.callsigns.indexOf(callsign) !== -1;
+    }
 
-        if (this.aircraft.callsigns.indexOf(callsign) === -1) {
-            this.aircraft.callsigns.push(callsign);
+    /**
+     * Add a new callsign to `aircraft.callsigns`
+     *
+     * @for AircraftController
+     * @method addCallsignToList
+     * @param callsign {string}
+     */
+    addCallsignToList(callsign) {
+        if (this.isCallsignInList(callsign)) {
+            // if you've made it here something has gone very wrong. generation of a callsign/flightNumber should
+            // also include verification that the callsign/flightNumber is unique
+            console.warn(`${callsign} already exists within the callsigns list!`);
+
+            return;
         }
 
-        return callsign;
+        this.aircraft.callsigns.push(callsign);
     }
 
     /**
@@ -359,6 +353,7 @@ export default class AircraftController {
         return this.aircraft.models[icao];
     }
 
+    // TODO: what is an `eid` and why would it beed to be updated?
     /**
      * Adjust all aircraft's eid values
      *